requirements.txt
alembic==1.9.2
attrs==21.4.0
autopep8==1.7.0
boto3==1.24.96
botocore==1.27.96
cachetools==5.2.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.4
cssselect==1.2.0
cssutils==2.6.0
Deprecated==1.2.14
emails==0.6
et-xmlfile==1.1.0
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Filter==0.1.0.dev5
Flask-JWT-Extended==4.3.1
Flask-Limiter==3.4.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
future==0.18.3
greenlet==2.0.1
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
jsonschema==4.7.2
limits==3.5.0
lxml==4.9.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.27.0
mdurl==0.1.2
mistune==2.0.4
numpy==1.23.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.0
pandas==1.4.3
premailer==3.10.0
psycopg2-binary==2.9.3
py-linq==1.4.0
pycodestyle==2.9.1
Pygments==2.16.1
PyJWT==2.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.5.2
s3transfer==0.6.0
Safe==0.4
six==1.16.0
SQLAlchemy==1.4.46
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.9
Werkzeug==2.2.2
wrapt==1.15.0
